+++ src/br/com/xkinfo/slc/service/ICondominioService.java
@@ -6,16 +6,16 @@ import java.util.ArrayList;
 
 public interface ICondominioService {
 
-    public void incluirCondominio(String nome, Integer cnpj,String endereco, Integer numero, String complemento, String bairro,
+    public void incluirCondominio(String nome, Float cnpj,String endereco, Float numero, String complemento, String bairro,
             String cidade, String estado, String sigla, String email, Usuario usuario) throws Exception;
 
-    public void alterarCondominio(Integer id, String nome, Integer cnpj,String endereco, Integer numero, String complemento, String bairro,
+    public void alterarCondominio(Integer id, String nome, Float cnpj,String endereco, Float numero, String complemento, String bairro,
             String cidade, String estado, String sigla, String email, Usuario usuario) throws Exception;
 
     public void excluirCondominio(Integer id) throws Exception;
 
     public ArrayList<Condominio> getCondominios() throws Exception;
 
-    public Condominio getCondominio(int id) throws Exception;
+    public Condominio getCondominio(Integer id) throws Exception;
 
 }
